+++ imageserver/sokdakimageserver/src/main/java/com/wooteco/sokdak/image/domain/Extension.java
@@ -1,5 +1,6 @@
-package com.wooteco.sokdak.image;
+package com.wooteco.sokdak.image.domain;
 
+import com.wooteco.sokdak.image.exception.NotSupportedExtensionException;
 import java.util.Arrays;
 
 public enum Extension {
